requirements.txt
# ============================================================
# ARES-7 v73 FULL - Python Dependencies
# ============================================================

# ========== Core Data Science ==========
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ========== Deep Learning ==========
torch>=2.0.0
flash-attn>=2.0.0
rotary-embedding-torch>=0.3.0

# ========== Machine Learning ==========
pytorch-tabnet>=4.0
catboost>=1.2
scikit-learn>=1.3.0
xgboost>=2.0.0

# ========== Broker Connectors ==========
mojito2>=0.3.0          # KIS (한국투자증권)
ib_insync>=0.9.86       # Interactive Brokers

# ========== Async & HTTP ==========
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0

# ========== Database ==========
sqlalchemy>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0  # PostgreSQL (optional)

# ========== Monitoring & Metrics ==========
prometheus-client>=0.19.0
dash>=2.14.0
plotly>=5.18.0

# ========== Configuration ==========
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.2

# ========== Utilities ==========
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0

# ========== Logging ==========
loguru>=0.7.0

# ========== Testing (optional) ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ========== Telegram Bot (optional) ==========
python-telegram-bot>=20.0

# ========== LLM APIs (optional) ==========
openai>=1.0.0
anthropic>=0.7.0

# ========== Financial Data ==========
yfinance>=0.2.0
alpha_vantage>=2.3.0

# ========== Type Hints ==========
typing-extensions>=4.8.0

# ========== Jupyter (development only) ==========
# jupyter>=1.0.0
# ipykernel>=6.25.0
